[DBZ-PGYB] Add load-balance property to multi host url pattern (#135)
Currently the PG debezium connector makes use of smart driver for
creating connections to YugabyteDB. The smart driver has in built
support for load balancing. However to enable this, the property
`load-balance=true` needs to be set in the connection URL.

This PR adds the `load-balance=true` property to the multi host URL used
for creating load-balanced connections.