[DBZ-PGYB][yugabyte/yugabyte-db#23082] Remove dependency on custom debezium-core changes (#137)
This PR reverts the changes made in `debezium-core` and converts those
changes to be in the local module only. Changes include:
1. Adding a custom field definition for `HOSTNAME` in
`PostgresConnectorConfig` which uses a new validator
`PostgresConnectorConfig#validateYBHostname`
a. Currently the validator definition returns a dummy positive result
but we can modify this in future to set a validation rule if needed.
2. Removing the method `SnapshotResult#isSkipped` and keeping it
localised in `PostgresSnapshotChangeEventSource`
3. Modifying `Dockerfile` to not copy custom `debezium-core-*.jar` -
this will enforce that the images use the prepackaged jar only.

This closes yugabyte/yugabyte-db#23082