DBZ-8100 Wait for streaming instead of starting connector
If we stop the connector short after its start, we need to wait for
streaming phase, i.e. wait for starting all connector tasks.
Waiting for connector start return `true` when connector begins to start
its tasks, but to prevent leaking of the resources, async engine doesn't
allow stopping the connector until all its tasks are started. Therefore
when using async. engine, we have to wait for tasks to be started before
stopping the connector, otherwise exception is thrown.