DBZ-8241 Use pre-defined builder factory is none is specified
Order of loading classes via SPI is not guaranteed and therefore the
builder factory returned by DebeziumEngine.create() and similar methods
is not deteministics. To avoid this confusion, when builder factory is
not specified, specify one as the default.
Until we remove `EmbeddedEngine`, `ConvertingEngineBuilderFactory` is
the default engine builder factory.