DBZ-8194 Always set binlog client GTID when GTID is enabled
When GTID is enabled, binlog client tracks GTID and eventually resumes
from the last GTID position. However, to enable GTID on the client,
client has to set initial GTID position (can be empty). When GTID is
enbled in the connector, always enable GTID also on the binlog client
to make sure that it tracks GTID and its keep alive thread eventually
resumes from the right position.