DBZ-8557: Implement keyset pagination in sqlserver connector
Sqlserver connector requires unbounded memory to process big
transactions. This happens due to the fact that `mssql-jdbc` driver
buffers any not yet consumed result sets in memory in order to execute
the next select statement. On the other hand the connector doesn't limit
the number of rows it selects. This leads to almost all transaction rows
being buffered in memory.

An obvious solution to that is to limit the number of rows selected.
This patch does exactly that. It introduces a new connector setting
called `streaming.fetch.size` (by analogy with `snapshot.fetch.size`) with the
default value of `0` which means the old behavior is preserved.