DBZ-2386 Notify errorHandler on Postgres errors
During certain scenarios, especially during RDS reboots (planned or
unplanned), Postgres becomes available in such a way that Debezium is
not aware that Postgres went away.  In many cases, this can result in
connectors thinking they are healthy (with all statuses returning
RUNNING), but are not actually reading from the replication slots.

By notifying the errorHandler when these occur, we are able to have the
in-built retry mechanism reconnect after these failures.

Due to the inconsistent nature of how this happens - not every reboot
will cause a connector to get stuck - we are catching erros happening in
both the PostgresStreamingChangeEventSource, and the
core ChangeEventSourceCoordinator.  In my testing, I have seen only some
types of events get caught by the error handler in the
ChangeEventSourceCoordinator.