DBZ-8467 Emit traces even when no propagated trace context is found
In the ActivateTracingSpan SMT, emitting traces when no propagated trace context is found is controlled by the
`tracing.with.context.field.only` setting. However, the if/else condition was flawed, resulting in traces not being
emitted event with this setting being set to `false`. This commit fixes the condition logic.