[DBZ-PGYB] Restrict retries when retry count reaches limit (#166)
## Problem

With the current retry model, it was being noticed that the connector
ended up retrying infinitely whenever an exception was thrown. This
could lead to false positives that the connector is still running while
the retries will keep failing.

## Solution

This PR addresses the issue by adding a check to the task layer so now
if the retry count reaches the maximum value, the connector will exit
and the task will reach in a failed state - this will help the end user
know the status of the task and act accordingly.

This PR also redefines the following properties and changes their
default values:
1. `errors.max.retries` - new default value is 60
2. `retriable.restart.connector.wait.ms` - new default value is 30000
(30s)

With the above change, the complete retry duration with the above
default configuration will now be 30 minutes. This effectively means
that if the connector/s task fails after exhausting all the retries then
it will go into a `FAILED` state.

For example, if the connector is needed to retry for a total of 30
minutes then we can handle it in 2 ways:
1. **By fixing the number of retries:** Let's say we want the number of
retries to be fixed to 15 so we can now configure our retry delay
accordingly i.e. `30 / 15 = 2 minutes = 120 s = 120000 ms`, so the
configuration will now add:

```json
"retriable.restart.connector.wait.ms":"120000",
"errors.max.retries":"15"
```

4. **By fixing the retry delay:** If we want to have a retry delay of a
minute then we will configure the number of retries accordingly i.e. `30
/ 1 = 30 retries`, so the configuration will now be:

```json
"retriable.restart.connector.wait.ms":"60000",
"errors.max.retries":"30"
```