DBZ-8651 Add libraries to override older Apicurio libs
Apicurio depends on older libraries than JDBC sink requires to work
properly. Libs which Apicurio depends on are part of Kafka distribution
but by including Apicurio and its libs, the older Apicurio libs have
precedence and JDBC sink fails to start. By including these libraries
into the connector we are trying to put the libs on the class path
before Apicurio libs so that the right ones are loaded for JDBC sink.