[#26004] Make load-balance property configurable (#171)
### Description
Currently the `load-balance` property in connection URL is hardcoded to
true for Multi host pattern. We do not mention any value to this
property for single host pattern and as a result it always defaults to
false.

This PR makes this property configurable via a connector config. A new
connector config `yb.load.balance.connections` has been introduced which
determines the value of `load-balance` property in the connection URL.
When set to true, the driver fetches and maintains a list of nodes from
the given endpoint (even when a single hostname is provided in the
config) in the universe and distributes the connections equally across
these nodes.

The default value of `yb.load.balance.connections` is true. We may want
to set this to false while testing scenarios where we want to have
imbalance of replication connections (walsenders) across nodes.

### Testing
Manually tested out by toggling the values of this config and providing
single host name in the connector config with a 3 node universe.