DBZ-8680 Postgres: Probe streaming connection during blocking snapshots
If there is no activity on the streaming connection while the blocking
snapshot is being taken, there is a possibility that the streaming
connection might be disconnected.

Periodically run a probing query on the streaming connection during the
blocking snapshot.

This is done by adding an overload of waitSnapshotCompletion which
allows the user to provide a heartbeat callback function, which is
regularly called while waiting.  This will allow other databases to
easily add the functionality as well.