[yugabyte/yugabyte-db#26069] Custom metrics implementation for multi task model (#165)
## Problem

Currently, the connector follows a single task single partition model
and thus uses the `DefaultChangeEventSourceMetricsFactory` which doesn't
let us override the metric object names. Since we cannot override names,
this causes an issue when we go for the multi task model as there's a
conflict in the metric object names if a task registers its metric
objects and subsequent tasks try to register their metric objects too.

## Solution

This is resolved by adding a custom metrics implementation which lets us
customise things accordingly for a multi task model. This is done by
following the example of metrics implemented by SqlServerConnector at
https://github.com/yugabyte/debezium/tree/2.5.2.Final/debezium-connector-sqlserver/src/main/java/io/debezium/connector/sqlserver/metrics

This closes yugabyte/yugabyte-db#26069