DBZ-8595 Remove unnecesarry refreshes of schema
Both `connectionCreated()` and `lockTablesForSchemaSnapshot()` are
called only from `RelationalSnapshotChangeEventSource#doExecute()`.
Before it we just determine snapshot tables, offset and do the locking
of the captured tables. We don't need schema for these operations and
it should be sufficient to refresh the schema only before taking the
actual snapshot. At this point fo time tables are already lock, so
we take schema refresh from consitent DB view.