DBZ-8595 Remove redundant schema refresh when starting streaming
Before this change we called schema refresh during initialization of
the streaming and once we started replicated connection just before
start of the streaming.

The first call during init has to be there as we may resume incremental
snapshot after streaming init step and need to have schema available.
The second refresh seems to be redundant as it's called almost right
after init and there doesn't seem to be anything which would require
to refresh the schema again just before start actual streaming.