[DBZ-PGYB] Specify only HYBRID_TIME when creating slot, skip otherwise (#175)
## Problem

With the introduction of `slot.lsn.type`, the default value we use for
creating the replication slot is `SEQUENCE` when the user does not
provide any value to the configuration property. However, while creating
the replication slot, if the user doesn't provide this config, the
`CREATE_REPLICATION_SLOT` statement looks like:

```
CREATE_REPLICATION_SLOT slot_name LOGICAL plugin_name SEQUENCE;
```

The above breaks the backward compatibility and causes the creation of
replication slot to fail on older versions where the newly added syntax
is not supported.

## Solution

As a fix, we will now only specify a LSN type while creating the
replication slot if it is `HYBRID_TIME`. For the case with `SEQUENCE`,
we will now rely on the default values as specified on the service to
create a replication slot of LSN type `SEQUENCE`.