[docs] Update Cassandra connector docs for conn config
This commit updates the Cassandra connector documentation for
connection configurations.
https://github.com/debezium/debezium-connector-cassandra/commit/3fe4963a61bf0919384f75267eb030f92ff15622#diff-cb8507d61cddae8b7d293cd7100ffe8afe87ef201c54a231d1a24243187f468a
removed many of the Cassandra connection configuration parameters and
replaced them with a single parameter, `cassandra.driver.config.file`,
which is a path to a configuration file for the Cassandra driver. The
removed connection parameters have to be configured with the Cassandra
driver now, not with the connector itself. This commit updates the
documentation to reflect this change. The language of the documentation
is mostly taken from this blog post,
https://debezium.io/blog/2022/03/25/debezium-1-9-cr1-released/, which
details the change.