Fix lint and formatting (#148)
## Problem
maven commands usually run formatting and linting by default. This leads
to some unknown files getting changed which leads to productivity loss.
Developer has to manually rollback those changes since they are always
unrelated to the change.

This PR is about committing those changes in the branch so that in
future only relevant lint and formatting changes are applied on using
maven commands.