DBZ-9000 Improve connector-level transaction filtering
When using LogMiner in buffered mode with connector-level filtering for username and
client id values, the connector would previously perform the filter at commit time
which lead to unnecessarily buffering the entire transaction to only discard it.

This change moves that filtering logic to the start of the transaction so that the
entire transaction can be discarded immediately and avoids adding any events for
that transaction to the buffer when excluded by username or client ids.

When using database-level filtering for added performance, those transactions are
never seen by the connector, thus never buffered regardless.