DBZ-9112 extend SQL parser to accept GRANT/REVOKE FLASHBACK
The old code recognized "FLASHBACK ARCHIVE" as an object privilege,
but "FLASHBACK" ("Issue a SQL Flashback Query on the table") without
the ARCHIVE part was ommitted. Flashback Queries ("AS OF") are
avaible for a very long time in Oracle Database - and in fact,
the Debezium Oracle Connector uses Flashback Queries when creating
snapshots.