Add mariadb grammar statements back to mysql grammar (#164)
## Problem
Although not supported, there are customers who are running mysql cdc v2
connector and connecting to a mariadb instance rather than a mysql
instance. However, since in v3.0.8.Final, the grammar file for mysql and
mariadb was seperated out, some of the ddl statements like `create or
replace table` which were being parsed correctly earlier would not be
passed.

[DBZ-8270](https://issues.redhat.com/browse/DBZ-8270) is the change
which went into.

## Testing
✅ UT/ITs pass
✅ Manual test done