DBZ-8747 Decouple mining range from fetch range
When the mining range and fetch range are identical, this can lead to
a variety of issues, including but not limited to seeing unsupported
operations due to transactions starting outside the mine window.

This change aims to decouple the mining and fetch ranges, where the
mining range will always start at the eldest transaction in the cache
will the fetch window will always increase based on batch windows.

This change also improves LOB handling as the same window logic will
now be used for when LOB is and isn't enabled, simplfying the logic.