DBZ-9306 Pause mining in archive log only mode when logs mined
The prior logic for mining archive logs in archive log only mode
did not account for all use cases that the connector could be in
when an archive log was fully mined.

This change uses heuristics to determine the appropriate point
when the connector should enter a wait state for a new archive
log to be created.