DBZ-9521 Flush commit queue on non-LOB tables
When LOB enabled tables are observed, their LOB data is stitched together
over the course of the transaction; however, when a table without an LOB
column is seen, such events are immediately dispatched. But this dispatch
would lead to event sequences being out of order.

This change is designed that if an immediate dispatch is necessary that
any pending changes in the queue are flushed by event order first and
only then will the non-LOB table be emitted, preserving event order.