Fix store state using JDBC documentation
I had to update this query because Oracle does not support a `LIMIT 1` clause. Since it didn't work I looked at the source and they specify a different property. I'm not sure if you consider the implementation or the documentation wrong at this point, but fixing the documentation is easier obviously. And backwards compatible.