DBZ-9680 Fix JDBC close/open behavior on log switch/max session time
The old behavior was slightly broken because even when the lower mining
boundary changed, the connection was closed and reopened when the config
contained log.mining.restart.connection, which was incorrect.

The JDBC connection should only be closed and reopened if and only if
the session max time was reached or if there was a log switch, but not
if the lower boundary changes.

Signed-off-by: Chris Cranford <chris@hibernate.org>