DBZ-9680 Rework logic to trigger when session SCN changes
Previously after each mining step, it was highly probable that the SCN
changed and this would trigger a LogMiner session restart. This has two
core impacts:

- Session reuse is limited
- When restart connection is enabled, it restarts on each mining step

This change aims to be smarter about triggering when the SCN changes by
limiting this toggle to only when the last and current batch values
point to different logs. This will allow the same session to be reused
while also preventing connection restarts on each mining step.

Signed-off-by: Chris Cranford <chris@hibernate.org>